+++ src/converter.h
@@ -24,11 +24,11 @@ enum class Status { Fail = -2, Pass = -1, Success = 0 };
 class Converter {
 public:
   virtual ~Converter() = default;
-  virtual Conv::Status convert(uint64_t event_id, char* in, size_t in_len,
-                               ForwardMode* pm) = 0;
+  virtual auto convert(uint64_t event_id, char* in, size_t in_len,
+                       ForwardMode* pm) -> Conv::Status = 0;
 
-  Matcher* get_matcher() { return matc.get(); }
-  virtual bool remaining_data() const { return false; }
+  auto get_matcher() -> Matcher* { return matc.get(); }
+  [[nodiscard]] virtual auto remaining_data() const -> bool { return false; }
   virtual void set_allowed_entropy_ratio(float e) {}
   void set_matcher(const std::string& filename, const Mode& mode);
   virtual void update_pack_message(uint64_t event_id, ForwardMode* pm,
@@ -67,13 +67,13 @@ public:
   PacketConverter(const uint32_t _l2_type);
   PacketConverter(const uint32_t _l2_type, time_t launch_time);
   PacketConverter(const PacketConverter&) = delete;
-  PacketConverter& operator=(const PacketConverter&) = delete;
+  auto operator=(const PacketConverter&) -> PacketConverter& = delete;
   PacketConverter(PacketConverter&&) = delete;
-  PacketConverter& operator=(const PacketConverter&&) = delete;
+  auto operator=(const PacketConverter &&) -> PacketConverter& = delete;
   ~PacketConverter() override = default;
-  Conv::Status convert(uint64_t event_id, char* in, size_t in_len,
-                       ForwardMode* msg) override;
-  bool remaining_data() const override
+  auto convert(uint64_t event_id, char* in, size_t in_len, ForwardMode* msg)
+      -> Conv::Status override;
+  auto remaining_data() const -> bool override
   {
     return sessions.get_number_bytes_in_sessions() > 0;
   }
@@ -85,8 +85,8 @@ public:
                            size_t max_bytes, const char* in = nullptr,
                            size_t in_len = 0) override;
 
-  Conv::Status payload_only_message(uint64_t event_id, ForwardMode* pm,
-                                    const char* in, size_t in_len);
+  auto payload_only_message(uint64_t event_id, ForwardMode* pm, const char* in,
+                            size_t in_len) -> Conv::Status;
 
 private:
   bool match;
@@ -106,21 +106,21 @@ private:
 
   void save_session(uint64_t event_id, uint32_t src, uint32_t dst,
                     uint8_t proto, uint16_t sport, uint16_t dport);
-  bool (PacketConverter::*get_l2_process())(unsigned char* offset,
-                                            size_t length);
-  bool (PacketConverter::*get_l3_process())(unsigned char* offset,
-                                            size_t length);
-  bool (PacketConverter::*get_l4_process())(unsigned char* offset,
-                                            size_t length);
-  bool l2_ethernet_process(unsigned char* offset, size_t length);
-  bool l2_null_process(unsigned char* offset, size_t length);
-  bool l3_ipv4_process(unsigned char* offset, size_t length);
-  bool l3_arp_process(unsigned char* offset, size_t length);
-  bool l3_null_process(unsigned char* offset, size_t length);
-  bool l4_icmp_process(unsigned char* offset, size_t length);
-  bool l4_udp_process(unsigned char* offset, size_t length);
-  bool l4_tcp_process(unsigned char* offset, size_t length);
-  bool l4_null_process(unsigned char* offset, size_t length);
+  auto (PacketConverter::*get_l2_process())(unsigned char* offset,
+                                            size_t length) -> bool;
+  auto (PacketConverter::*get_l3_process())(unsigned char* offset,
+                                            size_t length) -> bool;
+  auto (PacketConverter::*get_l4_process())(unsigned char* offset,
+                                            size_t length) -> bool;
+  auto l2_ethernet_process(unsigned char* offset, size_t length) -> bool;
+  auto l2_null_process(unsigned char* offset, size_t length) -> bool;
+  auto l3_ipv4_process(unsigned char* offset, size_t length) -> bool;
+  auto l3_arp_process(unsigned char* offset, size_t length) -> bool;
+  auto l3_null_process(unsigned char* offset, size_t length) -> bool;
+  auto l4_icmp_process(unsigned char* offset, size_t length) -> bool;
+  auto l4_udp_process(unsigned char* offset, size_t length) -> bool;
+  auto l4_tcp_process(unsigned char* offset, size_t length) -> bool;
+  auto l4_null_process(unsigned char* offset, size_t length) -> bool;
 };
 
 /**
@@ -131,12 +131,12 @@ class LogConverter : public Converter {
 public:
   LogConverter() = default;
   LogConverter(const LogConverter&) = delete;
-  LogConverter& operator=(const LogConverter&) = delete;
+  auto operator=(const LogConverter&) -> LogConverter& = delete;
   LogConverter(LogConverter&&) = delete;
-  LogConverter& operator=(const LogConverter&&) = delete;
+  auto operator=(const LogConverter &&) -> LogConverter& = delete;
   ~LogConverter() override = default;
-  Conv::Status convert(uint64_t event_id, char* in, size_t in_len,
-                       ForwardMode* msg) override;
+  auto convert(uint64_t event_id, char* in, size_t in_len, ForwardMode* msg)
+      -> Conv::Status override;
 };
 
 #endif