+++ src/util.cpp
@@ -6,18 +6,18 @@
 
 using namespace std;
 
-std::string& Util::ltrim(std::string& str)
+auto Util::ltrim(std::string& str) -> std::string&
 {
   str.erase(0, str.find_first_not_of("\t\n\v\f\r "));
   return str;
 }
 
-std::string& Util::rtrim(std::string& str)
+auto Util::rtrim(std::string& str) -> std::string&
 {
   str.erase(str.find_last_not_of("\t\n\v\f\r ") + 1);
   return str;
 }
 
-std::string& Util::trim(std::string& str) { return ltrim(rtrim(str)); }
+auto Util::trim(std::string& str) -> std::string& { return ltrim(rtrim(str)); }
 
 // vim: et:ts=2:sw=2